filter synced with charts
